requirements.txt
django == 2.2
python-dotenv == 0.10.3
psycopg2 == 2.8
djangorestframework == 3.10
django-filter == 2.2
markdown == 3.1
requests == 2.22
gunicorn == 20.0
sentry-sdk == 0.13.4
wheel == 0.33.6     # Needed for django-allauth
django-allauth==0.41.0
django-settings-export==1.2.1
django-crispy-forms==1.8.1
shortuuid==0.5.0
stripe==2.41.0
django-sendgrid-v5==0.8.1